resolving code review comments
